requirements.txt
easyocr==1.7.0

Flask==2.2.2
Flask_Cors==3.0.10

matplotlib==3.7.0
numpy==1.25.0


pandas==1.5.3

Pillow==9.4.0

pytesseract==0.3.10
tensorflow==2.10.0
tensorflow_object_detection_api==0.1.1
python-dotenv==1.0.0

